refactor: replace legacy runtime DB upgrades with baseline migrations
