+++ packages/server/tests/account-timezone-locale-defaults.test.ts
@@ -20,7 +20,22 @@ describe("account timezone/locale defaults", () => {
     expect(row.theme_preference).toBe("system");
   });
 
-  it("backfills null legacy account timezone/locale to system", () => {
+  it("adopts the schema version marker for already-current schema", () => {
+    const dir = mkdtempSync(join(tmpdir(), "everycal-db-"));
+    const dbPath = join(dir, "current.sqlite");
+    const initial = initDatabase(dbPath);
+    initial.pragma("user_version = 0");
+    initial.close();
+
+    const reopened = initDatabase(dbPath);
+    const userVersion = reopened.pragma("user_version", { simple: true }) as number;
+    expect(userVersion).toBe(1);
+
+    reopened.close();
+    rmSync(dir, { recursive: true, force: true });
+  });
+
+  it("rejects unsupported legacy account schemas", () => {
     const dir = mkdtempSync(join(tmpdir(), "everycal-db-"));
     const dbPath = join(dir, "legacy.sqlite");
     const legacy = new Database(dbPath);
@@ -32,150 +47,10 @@ describe("account timezone/locale defaults", () => {
         date_time_locale TEXT
       );
     `);
-    legacy.prepare("INSERT INTO accounts (id, username, timezone, date_time_locale) VALUES (?, ?, ?, ?)")
-      .run("u1", "user1", null, null);
     legacy.close();
 
-    const migrated = initDatabase(dbPath);
-    const row = migrated.prepare("SELECT timezone, date_time_locale FROM accounts WHERE id = ?").get("u1") as {
-      timezone: string;
-      date_time_locale: string;
-    };
+    expect(() => initDatabase(dbPath)).toThrow(/Unsupported legacy database schema/i);
 
-    expect(row.timezone).toBe("system");
-    expect(row.date_time_locale).toBe("system");
-
-    migrated.close();
-    rmSync(dir, { recursive: true, force: true });
-  });
-
-  it("preserves explicit legacy timezone/locale values", () => {
-    const dir = mkdtempSync(join(tmpdir(), "everycal-db-"));
-    const dbPath = join(dir, "legacy.sqlite");
-    const legacy = new Database(dbPath);
-    legacy.exec(`
-      CREATE TABLE accounts (
-        id TEXT PRIMARY KEY,
-        username TEXT NOT NULL UNIQUE,
-        timezone TEXT,
-        date_time_locale TEXT
-      );
-    `);
-    legacy.prepare("INSERT INTO accounts (id, username, timezone, date_time_locale) VALUES (?, ?, ?, ?)")
-      .run("u1", "user1", "Europe/Vienna", "en-GB");
-    legacy.close();
-
-    const migrated = initDatabase(dbPath);
-    const row = migrated.prepare("SELECT timezone, date_time_locale FROM accounts WHERE id = ?").get("u1") as {
-      timezone: string;
-      date_time_locale: string;
-    };
-
-    expect(row.timezone).toBe("Europe/Vienna");
-    expect(row.date_time_locale).toBe("en-GB");
-
-    migrated.close();
-    rmSync(dir, { recursive: true, force: true });
-  });
-
-  it("normalizes legacy column defaults for future inserts", () => {
-    const dir = mkdtempSync(join(tmpdir(), "everycal-db-"));
-    const dbPath = join(dir, "legacy.sqlite");
-    const legacy = new Database(dbPath);
-    legacy.exec(`
-      CREATE TABLE accounts (
-        id TEXT PRIMARY KEY,
-        username TEXT NOT NULL UNIQUE,
-        account_type TEXT NOT NULL DEFAULT 'person',
-        display_name TEXT,
-        bio TEXT,
-        avatar_url TEXT,
-        password_hash TEXT,
-        private_key TEXT,
-        public_key TEXT,
-        is_bot INTEGER NOT NULL DEFAULT 0,
-        discoverable INTEGER NOT NULL DEFAULT 0,
-        timezone TEXT NOT NULL DEFAULT 'Europe/Vienna',
-        date_time_locale TEXT NOT NULL DEFAULT 'en-GB',
-        default_event_visibility TEXT NOT NULL DEFAULT 'public',
-        created_at TEXT NOT NULL DEFAULT (datetime('now')),
-        updated_at TEXT NOT NULL DEFAULT (datetime('now')),
-        website TEXT,
-        city TEXT NOT NULL DEFAULT 'Wien',
-        city_lat REAL NOT NULL DEFAULT 48.2082,
-        city_lng REAL NOT NULL DEFAULT 16.3738,
-        email TEXT,
-        email_verified INTEGER NOT NULL DEFAULT 0,
-        email_verified_at TEXT,
-        preferred_language TEXT DEFAULT 'en'
-      );
-    `);
-    legacy.close();
-
-    const migrated = initDatabase(dbPath);
-    migrated.prepare("INSERT INTO accounts (id, username) VALUES (?, ?)").run("u2", "user2");
-    const row = migrated.prepare("SELECT timezone, date_time_locale FROM accounts WHERE id = ?").get("u2") as {
-      timezone: string;
-      date_time_locale: string;
-    };
-
-    expect(row.timezone).toBe("system");
-    expect(row.date_time_locale).toBe("system");
-
-    migrated.close();
-    rmSync(dir, { recursive: true, force: true });
-  });
-
-  it("preserves and normalizes theme_preference during legacy accounts rebuild", () => {
-    const dir = mkdtempSync(join(tmpdir(), "everycal-db-"));
-    const dbPath = join(dir, "legacy.sqlite");
-    const legacy = new Database(dbPath);
-    legacy.exec(`
-      CREATE TABLE accounts (
-        id TEXT PRIMARY KEY,
-        username TEXT NOT NULL UNIQUE,
-        account_type TEXT NOT NULL DEFAULT 'person',
-        display_name TEXT,
-        bio TEXT,
-        avatar_url TEXT,
-        password_hash TEXT,
-        private_key TEXT,
-        public_key TEXT,
-        is_bot INTEGER NOT NULL DEFAULT 0,
-        discoverable INTEGER NOT NULL DEFAULT 0,
-        timezone TEXT NOT NULL DEFAULT 'Europe/Vienna',
-        date_time_locale TEXT NOT NULL DEFAULT 'en-GB',
-        theme_preference TEXT NOT NULL DEFAULT 'system',
-        default_event_visibility TEXT NOT NULL DEFAULT 'public',
-        created_at TEXT NOT NULL DEFAULT (datetime('now')),
-        updated_at TEXT NOT NULL DEFAULT (datetime('now')),
-        website TEXT,
-        city TEXT NOT NULL DEFAULT 'Wien',
-        city_lat REAL NOT NULL DEFAULT 48.2082,
-        city_lng REAL NOT NULL DEFAULT 16.3738,
-        email TEXT,
-        email_verified INTEGER NOT NULL DEFAULT 0,
-        email_verified_at TEXT,
-        preferred_language TEXT DEFAULT 'en'
-      );
-    `);
-    legacy.prepare("INSERT INTO accounts (id, username, theme_preference) VALUES (?, ?, ?)")
-      .run("u1", "user1", "dark");
-    legacy.prepare("INSERT INTO accounts (id, username, theme_preference) VALUES (?, ?, ?)")
-      .run("u2", "user2", "AUTO");
-    legacy.close();
-
-    const migrated = initDatabase(dbPath);
-    const rows = migrated
-      .prepare("SELECT id, theme_preference FROM accounts WHERE id IN (?, ?) ORDER BY id")
-      .all("u1", "u2") as Array<{ id: string; theme_preference: string }>;
-
-    expect(rows).toEqual([
-      { id: "u1", theme_preference: "dark" },
-      { id: "u2", theme_preference: "system" },
-    ]);
-
-    migrated.close();
     rmSync(dir, { recursive: true, force: true });
   });
 });
